# #12155 Speed up twisted.web server, part 2