--- a/src/twisted/web/http.py
+++ b/src/twisted/web/http.py
@@ -2649,20 +2649,14 @@ class HTTPChannel(basic.LineReceiver, policies.TimeoutMixin):
         @param reason: The HTTP reason phrase to write.
         @type reason: L{bytes}
 
-        @param headers: The headers to write to the transport.
-        @type headers: L{twisted.web.http_headers.Headers}
+        @param headers: The headers to write to the transport, presumed to
+            already have been sanitized and deduplicated.
+        @type headers: Any iterable of two-tuples of L{bytes}, representing header
+            names and header values.
         """
-        sanitizedHeaders = Headers()
-        for name, value in headers:
-            sanitizedHeaders.addRawHeader(name, value)
-
         responseLine = version + b" " + code + b" " + reason + b"\r\n"
         headerSequence = [responseLine]
-        headerSequence.extend(
-            name + b": " + value + b"\r\n"
-            for name, values in sanitizedHeaders.getAllRawHeaders()
-            for value in values
-        )
+        headerSequence.extend(name + b": " + value + b"\r\n" for name, value in headers)
         headerSequence.append(b"\r\n")
         self.transport.writeSequence(headerSequence)