--- a/src/twisted/web/newsfragments/12155.feature
+++ b/src/twisted/web/newsfragments/12155.feature
@@ -1,0 +1,1 @@
+twisted.web's HTTP 1.1 server is an additional 5% faster.