--- a/src/twisted/web/test/requesthelper.py
+++ b/src/twisted/web/test/requesthelper.py
@@ -124,7 +124,13 @@ class DummyChannel:
         headerSequence.append(b"\r\n")
         self.transport.writeSequence(headerSequence)
 
-    writeHeadersPresanitized = writeHeaders
+    def writeHeadersObject(self, version, code, reason, headers):
+        self.writeHeaders(
+            version,
+            code,
+            reason,
+            [(k, v) for (k, values) in headers.getAllRawHeaders() for v in values],
+        )
 
     def getPeer(self):
         return self.transport.getPeer()