--- a/src/twisted/web/test/requesthelper.py
+++ b/src/twisted/web/test/requesthelper.py
@@ -118,19 +118,15 @@ class DummyChannel:
         pass
 
     def writeHeaders(self, version, code, reason, headers):
+        if isinstance(headers, Headers):
+            headers = [
+                (k, v) for (k, values) in headers.getAllRawHeaders() for v in values
+            ]
         response_line = version + b" " + code + b" " + reason + b"\r\n"
         headerSequence = [response_line]
         headerSequence.extend(name + b": " + value + b"\r\n" for name, value in headers)
         headerSequence.append(b"\r\n")
         self.transport.writeSequence(headerSequence)
-
-    def writeHeadersObject(self, version, code, reason, headers):
-        self.writeHeaders(
-            version,
-            code,
-            reason,
-            [(k, v) for (k, values) in headers.getAllRawHeaders() for v in values],
-        )
 
     def getPeer(self):
         return self.transport.getPeer()