@adiroiban 2024-05-03
Thanks for the update. Happy that this could be done without a new API. great work!